修改hdfs删除文件功能

+++ webapp/src/main/java/io/zeta/metaspace/web/task/quartz/QuartzJob.java
@@ -491,9 +491,11 @@ public class QuartzJob implements Job {
             return errorCount;
         } finally {
             checkResult(task, errorCount.floatValue(), 0f);
-            //删除hdfs中的文件，包括目录，实例：/apps/metaspace/project/griffin/0f11a928-37ee-402d-820d-8018acb9958f
-            //如果要调试代码的话，可以注释掉
-            hdfsUtils.deleteOnExit(LivyTaskSubmitHelper.getHdfsOutPathTaskId(task.getId()));
+            if(conf.getBoolean("metaspace.hdfs.file.delete.enable", true)){
+                //删除hdfs中的文件，包括目录，实例：/apps/metaspace/project/griffin/0f11a928-37ee-402d-820d-8018acb9958f
+                //如果要调试代码的话，可以注释掉
+                hdfsUtils.deleteOnExit(LivyTaskSubmitHelper.getHdfsOutPathTaskId(task.getId()));
+            }
         }
     }
